plugin should be the social zombie client and the social zombie interface is moved into lib
